Fix: Grid Card CSS 수정
