+++ src/components/mainpage/organisms/CardItem.tsx
@@ -1,9 +1,7 @@
-import { CardItemContainer } from "../../../styles/gridCardStyled/GridCardStyled";
 import Portrait from "../atoms/Portrait";
 import TitleArea from "../atoms/TitleArea";
 import IconArea from "../atoms/IconArea";
-import DateArea from "../atoms/DateArea";
-import { CardInfoContainer } from "../../../styles/gridCardStyled/GridCardStyled";
+import { CardInfoContainer, CardItemContainer } from "../../../styles/gridCardStyled/GridCardStyled";
 import { MainPageListType } from "../../../types/MainPageType";
 import { Link } from "react-router-dom";
 
@@ -11,10 +9,9 @@ const CardItem = (props: MainPageListType) => {
   return (
     <CardItemContainer thumbnail={props.thumbnail}>
       <Link to=''>
-        <CardInfoContainer>
-          <IconArea {...props}/>
-          <DateArea {...props}/>
+        <CardInfoContainer thumbnail={props.thumbnail}>
           <TitleArea {...props}/>
+          <IconArea {...props}/>
           <Portrait {...props}/>
         </CardInfoContainer>
       </Link>
