+++ src/components/mainpage/templates/GridCard.tsx
@@ -9,9 +9,7 @@ const GridCard = () => {
   return (
     <>
       <GridCardContainer>
-      {postList?.map((card) => (
-        <CardItem key={card.postId} {...card}/>
-      ))}
+      {postList?.map(card => <CardItem key={card.postId} {...card}/>)}
     </GridCardContainer>
     </>
   )
