+++ src/types/MainPageType.ts
@@ -1,4 +1,5 @@
 // 메인 리스트 타입
+// styled-components에서 불러서 쓰려고 ?:을 붙였습니다
 export interface MainPageListType {
   postId?: number;
   writerId?: number;
